+++ src/util.test.js
@@ -74,6 +74,7 @@ describe('convertToSlipData', () => {
       'stateProvRegion': 'region',
       'zipPostalCode': 'postalCode',
       'barcode': 'requester-barcode',
+      'departments': 'departments',
     },
     item: {
       'title': 'title',
@@ -129,6 +130,7 @@ describe('convertToSlipData', () => {
     expect(o['requester.zipPostalCode']).toEqual(source.requester.postalCode);
     expect(o['requester.barcode']).toEqual(source.requester.barcode);
     expect(o['requester.barcodeImage']).toEqual(`<Barcode>${source.requester.barcode}</Barcode>`);
+    expect(o['requester.departments']).toEqual(source.requester.departments);
 
     expect(o['item.title']).toEqual(source.item.title);
     expect(o['item.primaryContributor']).toEqual(source.item.primaryContributor);