+++ src/components/CheckInFooter/CheckInFooter.test.js
@@ -3,7 +3,7 @@ import {
   fireEvent,
   render,
   screen,
-} from '@testing-library/react';
+} from '@folio/jest-config-stripes/testing-library/react';
 
 import { runAxeTest } from '@folio/stripes-testing';