+++ test/bigtest/network/scenarios/default.js
@@ -1,6 +1,0 @@
-/* istanbul ignore file */
-
-export default function defaultScenario(server) {
-  server.create('staffSlip', { name: 'Hold' });
-  server.create('staffSlip', { name: 'Transit' });
-}