+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Also support `circulation` `14.0`. Refs UICHKIN-381.
 * Added requestDate token. Refs UICHKIN-384.
 * Add sessionId field to check-in request body JSON. Refs UICHKIN-385.
+* Upgrade babel config. Refs UICHKIN-389.
 
 ## [8.0.1] (https://github.com/folio-org/ui-checkin/tree/v8.0.1) (2023-03-28)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v8.0.0...v8.0.1)