+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added requestDate token. Refs UICHKIN-384.
 * Add sessionId field to check-in request body JSON. Refs UICHKIN-385.
 * Upgrade babel config. Refs UICHKIN-389.
+* Add possible for run axe tests. Refs UICHKIN-386.
 
 ## [8.0.1] (https://github.com/folio-org/ui-checkin/tree/v8.0.1) (2023-03-28)
 [Full Changelog](https://github.com/folio-org/ui-checkin/compare/v8.0.0...v8.0.1)